feat(ff-core, ff-backend-valkey): RFC-012 Stage 1a — EngineBackend trait + EngineError crate-move
Lands the Stage 1a abstraction layer per RFC-012 §5.1. Subsequent
stages (1b/c/d) migrate the hot paths through the trait; this PR is
additive to consumer-observable behaviour under the default feature.

- EngineError crate-move to ff-core::engine_error with a full
  ff-sdk::engine_error pub use shim. From<ScriptError> + the
  ScriptError-downcast helpers (valkey_kind, class, transport_script,
  transport_script_ref) live in ff-script::engine_error_ext because
  the orphan rule forbids defining those impls in ff-sdk (both types
  foreign) and the no-ff-core-on-ff-script direction is locked.
  ff-core exposes EngineError::class with a Transport->Terminal
  default; ff-script's engine_error_ext::class upgrades Transport
  to the Phase-1 classification when the inner source is a
  ScriptError. Consumers calling ff_sdk::SdkError::is_retryable /
  valkey_kind see no behaviour change.

- enrich_dependency_conflict converted from an inherent method on
  EngineError (now a foreign type to ff-sdk) to a free function in
  ff_sdk::engine_error. Zero external callers today.

- async-trait workspace dep at 0.1.

- EngineBackend trait in ff-core::engine_backend — 15 methods,
  Send + Sync + 'static, object-safe. _assert_dyn_compatible
  compile-time guard. RFC-012 §3.3 method set verbatim except
  complete(..., payload: Option<Vec<u8>>) diverges from the RFC
  sketch's Bytes (see Divergences below).

- BackendConfig + sub-types (BackendConnection, ValkeyConnection,
  BackendTimeouts, BackendRetry) in ff-core::backend. Stage 1a
  ships the Valkey arm only.

- FailOutcome moved to ff-core::backend::FailOutcome +
  ff_sdk::task::FailOutcome pub use shim. Existing consumers'
  exhaustive matches compile unchanged (ff-test,
  ff-readiness-tests — zero test edits).

- New crate ff-backend-valkey — sibling of ff-sdk.
  ValkeyBackend::connect(BackendConfig) -> Arc<dyn EngineBackend>,
  impl EngineBackend for ValkeyBackend. Hot-path methods return
  EngineError::Unavailable { op } per RFC-012 §5.2 staged-rollout
  contract; cancel_flow gets a real FCALL wrapper delegating to
  ff_script::functions::flow::ff_cancel_flow.

- ff-sdk additive API — FlowFabricWorker::connect_with(config,
  backend) (Stage 1a: stores the backend; still dials ferriskey
  internally for existing hot paths), backend(&self) ->
  Option<&Arc<dyn EngineBackend>> accessor.

- valkey-default cargo feature (default-on). Existing paths compile
  identically under the default feature. --no-default-features
  strips ferriskey from ff-sdk's direct deps and feature-gates
  task.rs, worker.rs, snapshot.rs, admin.rs, and the ferriskey-bound
  SdkError variants — agnosticism proof per RFC-012 §1.3.

- RFC-012 §3.3.0 BudgetId erratum — row annotated as already in tree.

- cargo build --workspace --all-features — green.
- cargo build -p ff-sdk --no-default-features — green (agnosticism proof).
- cargo test --workspace — green on ff-core, ff-script, ff-sdk,
  ff-backend-valkey, ff-engine, ff-scheduler, ff-server,
  ff-readiness-tests, ff-test. Zero existing test edits. Ferriskey
  shared-client integration tests are infra-bound.
- cargo clippy --workspace --all-targets -- -D warnings — green,
  including -p ff-sdk --no-default-features --all-targets.
- cargo semver-checks -p ff-core -p ff-script — 196 checks pass,
  no semver update required on both. ff-sdk local semver-checks
  blocked by aws-sdk toolchain requirement; CI resolves.

- complete(..., payload: Option<Vec<u8>>) vs RFC §3.3's Bytes.
  Matches the existing ff_sdk::ClaimedTask::complete signature for
  zero-churn consistency. Round-4 §7.17 resolved the payload
  container debate to Box<[u8]> for the public type (HandleOpaque);
  Option<Vec<u8>> is the zero-churn choice consistent with today's
  code and remains borrow-compatible via .as_deref().

- §3.4.1 suspend atomicity — preserved, not addressed. Out of
  Stage 1a scope.

- Stage 1b: ClaimedTask thin facade migration.
- Stage 1c: FlowFabricWorker hot-path forwarding + full
  BackendTimeouts/BackendRetry wiring.
- Stage 1d: snapshot.rs migration + .client() removal.

Refs: #89, RFC-012 §3.3, §3.3.0, §5.1.
