+++ crates/ff-core/src/lib.rs
@@ -2,6 +2,8 @@
 
 pub mod backend;
 pub mod contracts;
+pub mod engine_backend;
+pub mod engine_error;
 pub mod error;
 pub mod hash;
 pub mod keys;
@@ -9,3 +11,6 @@ pub mod partition;
 pub mod policy;
 pub mod state;
 pub mod types;
+
+// Convenience re-export so consumers can write `ff_core::EngineError`.
+pub use engine_error::EngineError;
